fix setup and teardown

+++ docker_hello_docker/.setup/setup.sh
@@ -2,7 +2,7 @@
 
 set -euxo pipefail
 
-if docker ps --format '{{.Names}}' | grep -w my-nginx &> /dev/null; then
+if docker ps -a --format '{{.Names}}' | grep -w my-nginx &> /dev/null; then
     docker rm -f my-nginx
 fi
